feat(nx-cloud): wip git utils
